+++ R/trackfreqs.R
@@ -8,7 +8,8 @@
 #'   xl = 1, osci = FALSE, gr = FALSE, sc = FALSE, bp = c(0, 22), cex = c(0.6, 1), 
 #'   threshold = 15, contour = "both", col = c("skyblue", "red2"),
 #'    pch = c(21, 24),  mar = 0.05, lpos = "topright", it = "jpeg", parallel = 1, 
-#'    path = NULL, img.suffix = NULL, custom.contour = NULL, pb = TRUE)
+#'    path = NULL, img.suffix = NULL, custom.contour = NULL, pb = TRUE, type = "p", 
+#'    leglab = c("Ffreq", "Dfreq"), col.alpha = 0.6)
 #' @param  X Data frame with results containing columns for sound file name (sound.files), 
 #' selection number (selec), and start and end time of signal (start and end).
 #' The ouptut of \code{\link{manualoc}} or \code{\link{autodetec}} can be used as the input data frame. 
@@ -81,6 +82,11 @@
 #' identical to the corresponding columns in X (same order). 
 #' @param pb Logical argument to control progress bar. Default is \code{TRUE}. Note that progress bar is only used
 #' when parallel = 1.
+#' @param type A character vector of length 1 indicating the type of frequency contour plot to be drawn. 
+#' Possible types are "p" for points, "l" for lines and "b" for both.
+#' @param leglab A character vector of length 1 or 2 containing the label(s) of the frequency contour legend 
+#' in the output image.
+#' @param col.alpha A numeric vector of length 1  within [0,1] indicating how transparent the lines/points should be.
 #' @return Spectrograms of the signals listed in the input data frame showing the location of 
 #' the dominant and fundamental frequencies.
 #' @family spectrogram creators
@@ -139,7 +145,8 @@ trackfreqs <- function(X, wl = 512, flim = c(0, 22), wn = "hanning", pal = rever
                        title = TRUE, propwidth = FALSE, xl = 1, osci = FALSE, gr = FALSE, sc = FALSE, 
                        bp = c(0, 22), cex = c(0.6, 1), threshold = 15, contour = "both", 
                        col = c("skyblue", "red2"),  pch = c(21, 24), mar = 0.05, lpos = "topright", 
-                       it = "jpeg", parallel = 1, path = NULL, img.suffix = NULL, custom.contour = NULL, pb = TRUE){     
+                       it = "jpeg", parallel = 1, path = NULL, img.suffix = NULL, custom.contour = NULL, pb = TRUE,
+                       type = "p", leglab = c("Ffreq", "Dfreq"), col.alpha = 0.6){     
   
   #check path to working directory
   if(!is.null(path))
@@ -177,6 +184,9 @@ trackfreqs <- function(X, wl = 512, flim = c(0, 22), wn = "hanning", pal = rever
   #wrap img creating function
   if(it == "jpeg") imgfun <- jpeg else imgfun <- tiff
   
+  #if type not l b or p
+  if(!any(type %in% c("p", "l", "b"))) stop(paste("Type", type, "not allowed"))  
+
   #join img.suffix and it
   if(is.null(img.suffix))
     img.suffix2 <- paste("trackfreqs", it, sep = ".") else   img.suffix2 <- paste(img.suffix, it, sep = ".")
@@ -213,12 +223,17 @@ trackfreqs <- function(X, wl = 512, flim = c(0, 22), wn = "hanning", pal = rever
       custom.contour <- custom.contour[match(paste(custom.contour[,c("sound.files")], custom.contour[,c("selec")]), paste(X[,c("sound.files")], X[,c("selec")])),]      
     }
   
-  #if only 1 pch was specfified
+  # adjust if only 1 pch was specfified
   if(length(pch) == 1) pch <- c(pch, pch)
+ 
+  # adjust if only 1 color was specified
+  if(length(col) == 1) col <- c(col, col)
+  
+  # adjust if only 1 leglab was specified
+  if(length(leglab) == 1) leglab <- c(leglab, leglab)
   
   #make colors transparent
-  if(length(col) == 1) col <- c(col, col)
-  col <- adjustcolor(c(col, "yellow", "black", "white", "red"), alpha.f = 0.7)
+  col <- adjustcolor(c(col, "yellow", "black", "white", "red"), alpha.f = col.alpha)
   
   
           trackfreFUN <- function(X, i, mar, flim, xl, picsize, wl, cexlab, inner.mar, outer.mar, res, bp, cex, threshold, pch, custom.contour){
@@ -228,6 +243,7 @@ trackfreqs <- function(X, wl = 512, flim = c(0, 22), wn = "hanning", pal = rever
     f <- r$sample.rate
     t <- c(X$start[i] - mar, X$end[i] + mar)
     
+    #adjust margins if signal is close to start or end of sound file
     mar1 <- mar
       
     if(t[1] < 0) {
@@ -291,12 +307,18 @@ if(contour %in% c("both", "ff") & is.null(custom.contour))
   ffreq <- matrix(ffreq[ffreq[,2] > b[1],], ncol = 2)
     
     # Plot extreme values fundamental frequency
-    points(c(ffreq[c(which.max(ffreq[,2]),which.min(ffreq[,2])),1]) + mar1, c(ffreq[c(which.max(ffreq[,2]), 
+      points(c(ffreq[c(which.max(ffreq[,2]),which.min(ffreq[,2])),1]) + mar1, c(ffreq[c(which.max(ffreq[,2]), 
         which.min(ffreq[,2])),2]), col = col[3], cex = cex[1] * 1.6, pch = pch[1], lwd = 2)  
   
   # Plot all fundamental frequency values
-    points(c(ffreq[,1])+mar1, c(ffreq[,2]), col = col[1], cex = cex[1], pch = pch[1], bg = col[1])
-    
+  if(type %in% c("p", "b"))
+     points(c(ffreq[,1])+mar1, c(ffreq[,2]), col = col[1], cex = cex[1], pch = pch[1], bg = col[1])
+ 
+  # plot lines      
+  if(type %in% c("l", "b"))
+    lines(ffreq[,1] + mar1, ffreq[,2], col = col[1], lwd = 3) 
+      
+         
     # Plot empty points at the bottom for the bins that did not detected any frequencies or out of bp
     if(nrow(ffreq1) > nrow(ffreq))
     points(c(ffreq1[!ffreq1[,1] %in% ffreq[,1], 1]) + mar1, rep(fl[1] + (fl[2] - fl[1]) * 0.04, nrow(ffreq1) - nrow(ffreq)), col = col[4], cex = cex[1] * 0.7, pch = pch[1])
@@ -311,16 +333,20 @@ if(contour %in% c("both", "ff") & is.null(custom.contour))
       dfreq <- matrix(dfreq1[!is.na(dfreq1[,2]),], ncol = 2)  
 
     # Plot extreme values dominant frequency
-    points(c(dfreq[c(which.max(dfreq[,2]),which.min(dfreq[,2])),1])+mar1, c(dfreq[c(which.max(dfreq[,2]),
+        points(c(dfreq[c(which.max(dfreq[,2]),which.min(dfreq[,2])),1])+mar1, c(dfreq[c(which.max(dfreq[,2]),
         which.min(dfreq[,2])),2]), col = col[3], cex = cex[1] * 1.6, pch = pch[2], lwd = 2) 
 
    # Plot all dominant frequency values
-    points(dfreq[,1] + mar1, dfreq[,2], col = col[2], cex = cex[1], pch = pch[2], bg = col[2]) 
+   if(type %in% c("p", "b"))
+     points(dfreq[,1] + mar1, dfreq[,2], col = col[2], cex = cex[1], pch = pch[2], bg = col[2]) 
+    
+  # plot lines      
+    if(type %in% c("l", "b"))
+    lines(dfreq[,1] + mar1, dfreq[,2], col = col[2], lwd = 3) 
     
     # Plot empty points at the bottom for the bins that did not detected any frequencies or out of bp
     if(nrow(dfreq1) > nrow(dfreq))
       points(c(dfreq1[!dfreq1[,1] %in% dfreq[,1], 1]) + mar1, rep(fl[1] + (fl[2] - fl[1]) * 0.02, nrow(dfreq1) - nrow(dfreq)), col = col[4], cex = cex[1] * 0.7, pch = pch[2])
-    
         }
     
     # Use freq values provided by user   
@@ -336,35 +362,48 @@ if(contour %in% c("both", "ff") & is.null(custom.contour))
                                                                                       which.min(freq[,2])),2]), col = col[3], cex = cex[1] * 1.6, pch = pch[2], lwd = 2) 
       
       # Plot all dominant frequency values
-      points(freq[,1] + mar1, freq[,2], col = col[2], cex = cex[1], pch = pch[1], bg = col[2]) 
+      if(type %in% c("p", "b"))
+        points(freq[,1] + mar1, freq[,2], col = col[2], cex = cex[1], pch = pch[2], bg = col[2]) 
+      
+      
+      # plot lines      
+      if(type %in% c("l", "b"))
+        lines(freq[,1] + mar1, freq[,2], col = col[2], lwd = 3) 
       
       # Plot empty points at the bottom for the bins that did not detected any frequencies or out of bp
       if(nrow(freq1) > nrow(freq))
-        points(c(freq1[!freq1[,1] %in% freq[,1], 1]) + mar1, rep(fl[1] + (fl[2] - fl[1]) * 0.02, nrow(freq1) - nrow(freq)), col = col[4], cex = cex[1] * 0.7, pch = pch[1])
+        points(c(freq1[!freq1[,1] %in% freq[,1], 1]) + mar1, rep(fl[1] + (fl[2] - fl[1]) * 0.02, nrow(freq1) - nrow(freq)), col = col[4], cex = cex[1] * 0.7, pch = pch[2])
       
     }
     
     abline(v = c(mar1, mar2), col= col[6], lty = "dashed")
     
-    # Legend coordinates can be uniquely adjusted 
-  if(is.null(custom.contour))
-     { if(contour == "both")
-    legend(lpos, legend = c("Ffreq", "Dfreq"), bg = col[5],
-           pch = pch, col = col[1:2], bty = "o", cex = cex[2], pt.bg = col[1:2])
+## legend
+      # remove points fo legend
+    if(type == "l") pch <- NA
+    if(type %in% c("l", "b")) lwd <- 3 else lwd = NA
+    
+    # Adjust legend coordinates  
+    if(is.null(custom.contour))
+    {
+      if(contour == "both")
+    legend(lpos, legend = leglab, bg = col[5],
+           pch = pch, col = col[1:2], bty = "o", cex = cex[2], pt.bg = col[1:2], lwd = lwd)
 
     if(contour == "ff")
-      legend(lpos, legend = "Ffreq",
-             pch = pch[1], col = col[1], bty = "o", cex = cex[2], bg = col[5], pt.bg = col[1])
+      legend(lpos, legend = leglab[1],
+             pch = pch[1], col = col[1], bty = "o", cex = cex[2], bg = col[5], pt.bg = col[1], lwd = lwd)
 
     if(contour == "df")
-      legend(lpos, legend = "Dfreq",
-             pch = pch[2], col = col[2], bty = "o", cex = cex[2], bg = col[5], pt.bg = col[2])
-    } else
-               legend(lpos, legend = "Freq",
-                      pch = pch[1], col = col[1], bty = "o", cex = cex[2], bg = col[5], pt.bg = col[1])
+      legend(lpos, legend = leglab[2],
+             pch = pch[2], col = col[2], bty = "o", cex = cex[2], bg = col[5], pt.bg = col[2], lwd = lwd)
+      } else
     
+{ 
+legend(lpos, legend = leglab[1],
+             pch = pch[2], col = col[2], bty = "o", cex = cex[2], bg = col[5], pt.bg = col[2], lwd = lwd)}
     
-    invisible() # execute par(old.par)  
+    invisible() 
     dev.off()
     return(NULL)